Add VictoriaMetrix as InfluxDB alternative
The fate of InfuxDB is unclear with many home users not switching to version 3 of InfluxDB. Being stuck at an older version brings with it the danger of unfixed security issues, but there is hope. VictoriaMetrics has among others a REST API, which is compatible with InfluxDB. Thus it can be used as a drop in replacement in evcc.

However Grafana dashboards have to be rebuild, because it does not support any of InfluxDBs query languages and users internally a different data model.